refactor: tidying up
**Restructure**

- combine & rename Type files

**Notes & Comments**

- `api/neo4j/writeBeacon.ts`
- `routes/hubRoutes.ts`
- `utils/language/breaker.ts`
