--- types/security.ts
@@ -1,5 +1,0 @@
-export interface CredsN4J {
-  URI: string;
-  USER: string;
-  PASSWORD: string;
-}
