+++ types/server.ts
@@ -4,3 +4,9 @@ export interface Subrouter {
   router: Router;
   routes: string[]
 }
+
+export interface CredsN4J {
+  URI: string;
+  USER: string;
+  PASSWORD: string;
+}
